Name added to High Scores
